changed blink from float to long

+++ src/main/java/eu/eutampieri/catacombs/ui/EndGameState.java
@@ -15,12 +15,12 @@ import java.awt.event.KeyEvent;
 
 public class EndGameState extends State {
 
-    private static final float BLINK_DELAY = 500;
+    private static final long BLINK_DELAY = 500;
     /**
      * Font used to write the end game message.
      */
     private final Font font = new Font("Monospace", Font.PLAIN, 40);
-    private float blinkDelayCount;
+    private long blinkDelayCount;
     private boolean blink = true;
     private final DungeonGame game;
